refactor(ai/test): split diplomatic_candidate_scoring_test into focused files
Splits `packages/colonizethis_ai/test/diplomatic_candidate_scoring_test.dart`
(616 lines, 3 groups) into three focused test files mapped to the
post-`#2289` AI source layout under `lib/src/planning/`:

- `diplomatic_candidate_scoring_test.dart` (233 lines) keeps the
  `computeDiplomaticCandidateScores` group only — exercises the pure
  scoring helper in `planning/diplomatic_candidate_scoring.dart`.
- `diplomacy_planner_cooldowns_test.dart` (302 lines, new) holds the
  `diplomacy planner cooldowns` group — exercises wardec/improve-relations
  cooldowns through `computeDiplomaticCandidateScores` plus
  `runDomainPlanners` (`planning/diplomacy_planner.dart` +
  `planning/domain_planner_orchestrator.dart`).
- `move_planner_diplomacy_filter_test.dart` (98 lines, new) holds the
  `move planner diplomacy filter` test — exercises the move planner's
  at-war vs at-peace destination weighting through `runDomainPlanners`
  (`planning/move_planner.dart`).

No test bodies were edited; only the surrounding `main()` wrappers and
imports were duplicated into each new file. The new files import the
same shared fakes from `domain_planner_test_fake_api.dart`. The scoring
file no longer needs that fake import.

Verification:
- `dart analyze` for `packages/colonizethis_ai` — clean.
- `dart test` for `packages/colonizethis_ai` — 124 tests pass (same
  count as before; the 7 tests previously in the merged file now run as
  2 + 4 + 1 across the three split files).

Refs #2289
